Refactor: 커스텀 타입들 별도의 파일로 분리
