+++ src/components/Header.tsx
@@ -5,13 +5,7 @@ import { faChevronLeft } from '@fortawesome/free-solid-svg-icons';
 import { faHome } from '@fortawesome/free-solid-svg-icons';
 import { faGear } from '@fortawesome/free-solid-svg-icons';
 import { useRouter } from 'next/navigation';
-
-type HeaderProps = {
-  goHome: boolean;
-  canSet: boolean;
-  onSet?: () => void;
-  title: string;
-};
+import { HeaderProps } from '@/types/global';
 
 export default function Header({ goHome, canSet, onSet, title }: HeaderProps) {
   const router = useRouter();
