+++ src/components/Option.tsx
@@ -2,14 +2,8 @@ import '@fortawesome/fontawesome-svg-core/styles.css';
 import { config } from '@fortawesome/fontawesome-svg-core';
 import Card from './Card';
 import { ReactElement } from 'react';
+import { OptionProps } from '@/types/global';
 config.autoAddCss = false; // 자동으로 CSS를 추가하지 않도록 설정
-
-type OptionProps = {
-  color: string;
-  title?: string;
-  content?: string;
-  icon?: ReactElement;
-};
 
 export default function Option({ color, title, content, icon }: OptionProps) {
   return (
