--- src/utils/teamTypes.ts
@@ -1,39 +1,0 @@
-export enum TeamType {
-  'Animals',
-  'Colors',
-  'Numbers',
-  'Countries',
-}
-
-export function animalTeams() {
-  return [
-    ['🦝', '너구리'],
-    ['🐮', '젖소'],
-    ['🦁', '사자'],
-    ['🐯', '호랑이'],
-    ['🐨', '코알라'],
-    ['🐼', '판다'],
-    ['🐻‍❄️', '북극곰'],
-    ['🐻', '불곰'],
-    ['🐰', '토끼'],
-    ['🐹', '햄스터'],
-    ['🐭', '쥐'],
-    ['🐱', '고양이'],
-    ['🐶', '강아지'],
-    ['🐷', '돼지'],
-    ['🐴', '말'],
-    ['🦊', '여우'],
-    ['🐺', '늑대'],
-    ['🐥', '병아리'],
-    ['🐧', '펭귄'],
-    ['🐵', '원숭이'],
-    ['🐸', '개구리'],
-    ['🐋', '고래'],
-    ['🦈', '상어'],
-    ['🐟', '물고기'],
-    ['🐙', '문어'],
-    ['🦉', '부엉이'],
-    ['🐓', '닭'],
-    ['🦅', '독수리'],
-  ];
-}
